requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.7.6
amqp==2.6.1
async-timeout==4.0.2
attrs==21.4.0
authy==2.2.6
Babel==2.9.1
base58==2.1.1
billiard==3.6.4.0
billplz-python==1.0.2
bitarray==1.2.2
CacheControl==0.12.10
cachetools==5.0.0
celery==4.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cytoolz==0.11.2
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
firebase-admin==5.2.0
Flask==2.0.3
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-ReCaptcha==0.4.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
flower==0.9.3
frozenlist==1.3.0
google-api-core==2.6.0
google-api-python-client==2.39.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-cloud-core==2.2.2
google-cloud-firestore==2.3.4
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.3.1
googleapis-common-protos==1.55.0
greenlet==1.1.2
grpcio==1.44.0
grpcio-status==1.44.0
gunicorn==20.1.0
hexbytes==0.2.2
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.1
importlib-resources==5.4.0
ipfshttpclient==0.8.0a2
itsdangerous==2.1.0
Jinja2==3.0.3
jsonschema==3.2.0
kombu==4.6.4
lru-dict==1.1.7
Mako==1.1.6
MarkupSafe==2.1.0
msgpack==1.0.3
multiaddr==0.0.9
multidict==6.0.2
mysqlclient==2.1.0
netaddr==0.8.0
packaging==21.3
parsimonious==0.8.1
proto-plus==1.20.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.14.1
pyparsing==3.0.7
pyrsistent==0.18.1
python-decouple==3.6
pytz==2021.3
redis==3.2.1
requests==2.27.1
rlp==2.0.1
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.31
stripe==2.66.0
toolz==0.11.2
tornado==5.1.1
uritemplate==4.1.1
urllib3==1.26.8
varint==1.0.2
vine==1.3.0
web3==5.28.0
websockets==9.1
Werkzeug==2.0.3
WTForms==3.0.1
yarl==1.7.2
zipp==3.7.0
